requirements.txt
h5py==2.7.1
matplotlib==3.0.3
numpy==1.17.0
pandas==0.24.2
scikit_learn==0.19.1
spacy==2.0.8
tensorflow-gpu==1.13.2
tqdm==4.19.6
scipy==1.3.0
tensorflow-hub
tensorflow_ranking